Move `app` under `src`
